Moves some command files to the right place
