+++ fabfile.py
@@ -3,7 +3,7 @@ from invoke.config import Config
 from fabric import task
 
 from environments.surfpol import create_configuration_and_session
-from postgres.tasks_remote import restore_snapshot, setup_postgres, create_snapshot
+from commands.postgres.fabric import restore_snapshot, setup_postgres, create_snapshot
 from commands.projects.harvester.fabric import create_super_user, connect_uwsgi, connect_flower
 from legacy import download_database
