+++ tasks.py
@@ -4,7 +4,7 @@ from environments.surfpol import create_configuration_and_session
 from elastic.tasks import setup, create_snapshot, load_repository, restore_snapshot
 from deploy import prepare_builds, build, push, deploy, migrate
 from test import e2e_tests, harvester_tests, service_tests, elastic_search_tests, test
-from postgres.tasks_local import import_snapshot as postgres_import_snapshot
+from commands.postgres.invoke import import_snapshot as postgres_import_snapshot
 from commands.projects.harvester.invoke import setup_harvester, import_dataset, harvest, cleanup
 from service.tasks_local import setup_service
 from legacy import upload_database, download_media, upload_media
